backend/requirements.txt
Django==2.2.19
pytz==2022.7.1
sqlparse==0.4.3

django-colorfield==0.8.0
Pillow==9.4.0

yapf==0.32.0
flake8==6.0.0

djangorestframework==3.12.3

djoser==2.1.0
djangorestframework-simplejwt==4.7.2
